backend/pyproject.toml
[project]
name = "talkam-backend"
version = "0.1.0"
description = "FastAPI backend scaffold for Talkam Liberia"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic[email]>=2.9.0",
  "pydantic-settings>=2.5.0",
  "eval-type-backport>=0.2.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "asyncpg>=0.29.0",
  "alembic>=1.13.2",
  "redis>=5.2.0",
  "aioboto3>=13.1.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "httpx>=0.27.0",
  "orjson>=3.10.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.8",
  "pytest-cov>=4.1.0",
  "ruff>=0.6.9",
  "pre-commit>=3.8.0"
]

[tool.uv]
dev-dependencies = ["talkam-backend[dev]"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E","F","I","N","UP","S","ASYNC"]

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
